ffmpeg/qsv: move AVC Encoder Test classes to lib
...so they can be shared.
